Fix some exceptions
Catch errors on some queries for defunct accessibles.  Note that many calls
on defunct accessibles throw exceptions.
Also, when receiving a ChildrenChanged and the list is not large enough to hold
the referenced child, refresh the children.  This works around ChildrenChanged
signals that we missed because we had not yet registered a match rule, for
instance.  We really need to solve this in managed-dbus somehow--there is no
reason not to register a dbus match-rule for ChildrenChanged and some other
signals with no sender/path except that managed-dbus does not support it
without the use of internal classes.